LPD-56968 Rename and move package
